requirements.txt
ultralytics
torch
onnx
onnxruntime
ipython
cv2
numpy
matplotlib
Pillow
argparse
